// src/app/models/subscribers.interface.ts
export interface SubscribersI {
    Name: string;
    Activity: string;
    Area: string;
    ConnectionState: number;
    CountryCode: string;
    CountryName: string;
    Email: string;
    EndpointsCount: number;
    Id: any;
    JobTitle: string;
    LastActivity: null;
    LastActivityString: null;
    LastActivityUtc: null;
    PhoneCode: string;
    PhoneCodeAndNumber: string;
    PhoneNumber: string;
    PublicId:number;
    SubscriptionDate: null;
    SubscriptionMethod: number;
    SubscriptionState: number;
    SubscriptionStateDescription: string;
    SystemId: null;
    Topics:any;
    ValidEmail: boolean;
}